Convert another controller test to feature test
Decouple from implementation, test closer to HTTP...
